update server port
